requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
celery
redis
supabase
groq
langchain-core
pdf2image
Pillow
pytesseract
sentence-transformers
faiss-cpu